Extract EMF interpretation helpers
